🔨 #44 플레이리스트 내 상세 비디오 데이터를 가져오는 로직 hooks로 분리
